Separate gimbal guns into turret and targeting
