+++ Content.Server/ShipGuns/Components/TurretConsoleComponent.cs
@@ -8,9 +8,5 @@ namespace Content.Server.ShipGuns.Components;
 [ComponentReference(typeof(SharedTurretConsoleComponent))]
 public sealed class TurretConsoleComponent : SharedTurretConsoleComponent
 {
-    /// <summary>
-    /// How much to zoom out when using ship guns.
-    /// </summary>
-    [DataField("zoom")]
-    public Vector2 Zoom = new(1.5f, 1.5f);
+    
 }
