+++ Content.Shared/ShipGuns/Components/GimbalGunManualComponent.cs
@@ -1,0 +1,10 @@
+﻿namespace Content.Shared.ShipGuns.Components;
+
+/// <summary>
+/// For gimbal guns that have to be controlled by a human.
+/// </summary>
+[RegisterComponent]
+public sealed class GimbalGunManualComponent : Component
+{
+
+}
