+++ Content.Shared/ShipGuns/Components/GimbalGunTurretComponent.cs
@@ -1,0 +1,10 @@
+﻿namespace Content.Shared.ShipGuns.Components;
+
+/// <summary>
+/// For gimbal guns that automatically target things (aka turrets)
+/// </summary>
+[RegisterComponent]
+public sealed class GimbalGunTurretComponent : Component
+{
+
+}
