+++ Content.Shared/ShipGuns/Components/SharedTurretConsoleComponent.cs
@@ -1,5 +1,4 @@
 using Robust.Shared.GameStates;
-using Robust.Shared.Serialization;
 
 namespace Content.Shared.ShipGuns.Components;
 
@@ -9,18 +8,5 @@ namespace Content.Shared.ShipGuns.Components;
 [NetworkedComponent]
 public abstract class SharedTurretConsoleComponent : Component
 {
-    // Only one because tracking more than one mouse sounds fucking awful.
-    /// <summary>
-    /// Who's using the turret console at the moment
-    /// </summary>
-    [ViewVariables]
-    public GunnerComponent? SubscribedGunner = null;
 
-    public GimbalGunTargetComponent? Target;
-}
-
-[Serializable, NetSerializable]
-public enum TurretConsoleUiKey : byte
-{
-    Key,
 }
