+++ Resources/Prototypes/Entities/Structures/Machines/Computers/computers.yml
@@ -629,7 +629,7 @@
   parent: BaseComputer
   id: BaseComputerTurret
   name: turret console
-  description: Used to control ship guns.
+  description: Used to control turrets.
   abstract: true
   components:
     - type: TurretConsole
@@ -640,7 +640,6 @@
       interfaces:
         - key: enum.ShuttleConsoleUiKey.Key
           type: ShuttleConsoleBoundUserInterface
-    - type: RadarConsole
     - type: ExtensionCableReceiver
     - type: PointLight
       radius: 1.5
@@ -653,6 +652,41 @@
   parent: BaseComputerTurret
   id: ComputerTurret
   name: turret console
+  description: Used to control turrets.
+  components:
+    - type: Appearance
+      visuals:
+        - type: ComputerVisualizer
+          key: generic_key
+          screen: shuttle
+
+- type: entity
+  parent: BaseComputer
+  id: BaseComputerTargeting
+  name: targeting console
+  description: Used to control ship guns.
+  abstract: true
+  components:
+    - type: TargetingConsole
+    - type: ActivatableUI
+      key: enum.ShuttleConsoleUiKey.Key
+    - type: ActivatableUIRequiresPower
+    - type: UserInterface
+      interfaces:
+        - key: enum.ShuttleConsoleUiKey.Key
+          type: ShuttleConsoleBoundUserInterface
+    - type: ExtensionCableReceiver
+    - type: PointLight
+      radius: 1.5
+      energy: 1.6
+      color: "#43ccb5"
+    - type: Rotatable
+      rotateWhileAnchored: true
+
+- type: entity
+  parent: BaseComputerTargeting
+  id: ComputerTargeting
+  name: targeting console
   description: Used to control ship guns.
   components:
     - type: Appearance
